+++ src/main/java/rbasamoyai/createbigcannons/munitions/big_cannon/BigCannonProjectileBlockEntity.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import com.simibubi.create.api.equipment.goggles.IHaveGoggleInformation;
 import com.simibubi.create.api.schematic.nbt.PartialSafeNBT;
+import com.simibubi.create.api.schematic.requirement.SpecialBlockEntityItemRequirement;
+import com.simibubi.create.content.schematics.requirement.ItemRequirement;
 import com.simibubi.create.foundation.blockEntity.SyncedBlockEntity;
 import com.simibubi.create.foundation.utility.CreateLang;
 
@@ -21,8 +23,10 @@ import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import rbasamoyai.createbigcannons.index.CBCDataComponents;
 import rbasamoyai.createbigcannons.index.CBCItems;
+import rbasamoyai.createbigcannons.utils.CBCUtils;
 
-public class BigCannonProjectileBlockEntity extends SyncedBlockEntity implements IHaveGoggleInformation, Container, PartialSafeNBT {
+public class BigCannonProjectileBlockEntity extends SyncedBlockEntity implements IHaveGoggleInformation, Container, PartialSafeNBT,
+    SpecialBlockEntityItemRequirement {
 
 	public BigCannonProjectileBlockEntity(BlockEntityType<?> type, BlockPos pos, BlockState state) {
 		super(type, pos, state);
@@ -43,16 +47,23 @@ public class BigCannonProjectileBlockEntity extends SyncedBlockEntity implements
 
     @Override
     public void writeSafe(CompoundTag tag, HolderLookup.Provider registries) {
-        PatchedDataComponentMap restore = new PatchedDataComponentMap(this.components());
         PatchedDataComponentMap copy = new PatchedDataComponentMap(DataComponentMap.EMPTY);
         this.writeSafeComponents(copy);
-        this.setComponents(copy);
-        super.saveAdditional(tag, registries);
-        this.setComponents(restore);
+        this.saveAdditional(tag, registries);
+        CBCUtils.saveComponentsToStructureTag(tag, copy, registries);
     }
 
     protected void writeSafeComponents(PatchedDataComponentMap safeComponents) {
+        ItemStack tracer = this.getTracer();
+        if (tracer.isEmpty()) {
+            safeComponents.set(CBCDataComponents.TRACER, ItemContainerContents.EMPTY);
+        } else {
+            safeComponents.set(CBCDataComponents.TRACER, ItemContainerContents.fromItems(List.of(tracer)));
+        }
+    }
 
+    protected void writeSafeComponentsForItemRequirement(PatchedDataComponentMap safeComponents) {
+        this.writeSafeComponents(safeComponents);
     }
 
     @Override
@@ -76,13 +87,13 @@ public class BigCannonProjectileBlockEntity extends SyncedBlockEntity implements
 	}
 
     public void setTracer(ItemStack itemStack) {
-        PatchedDataComponentMap components = new PatchedDataComponentMap(this.components());
+        PatchedDataComponentMap components = new PatchedDataComponentMap(DataComponentMap.EMPTY);
         if (itemStack.isEmpty()) {
             components.remove(CBCDataComponents.TRACER);
         } else {
             components.set(CBCDataComponents.TRACER, ItemContainerContents.fromItems(List.of(itemStack)));
         }
-        this.setComponents(components);
+        this.applyComponents(this.components(), components.asPatch());
     }
 
 	@Override
@@ -137,4 +148,16 @@ public class BigCannonProjectileBlockEntity extends SyncedBlockEntity implements
 			.forGoggles(tooltip);
 		return true;
 	}
+
+    @Override
+    public ItemRequirement getRequiredItems(BlockState state) {
+        ItemStack stack = new ItemStack(this.getBlockState().getBlock().asItem());
+        if (stack.isEmpty())
+            return ItemRequirement.INVALID;
+        PatchedDataComponentMap safeComponents = new PatchedDataComponentMap(DataComponentMap.EMPTY);
+        this.writeSafeComponentsForItemRequirement(safeComponents);
+        stack.applyComponents(safeComponents);
+        return new ItemRequirement(new ItemRequirement.StrictNbtStackRequirement(stack, ItemRequirement.ItemUseType.CONSUME));
+    }
+
 }