+++ src/main/java/rbasamoyai/createbigcannons/munitions/big_cannon/FuzedBlockEntity.java
@@ -8,6 +8,7 @@ import com.simibubi.create.foundation.utility.CreateLang;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderLookup;
+import net.minecraft.core.component.DataComponentMap;
 import net.minecraft.core.component.PatchedDataComponentMap;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
@@ -37,6 +38,17 @@ public class FuzedBlockEntity extends BigCannonProjectileBlockEntity {
     public void readClient(CompoundTag tag, HolderLookup.Provider registries) {
         super.readClient(tag, registries);
         this.setFuze(ItemStack.parseOptional(registries, tag.getCompound("Fuze")));
+    }
+
+    @Override
+    protected void writeSafeComponents(PatchedDataComponentMap safeComponents) {
+        super.writeSafeComponents(safeComponents);
+        ItemStack fuze = this.getFuze();
+        if (fuze.isEmpty()) {
+            safeComponents.set(CBCDataComponents.FUZE, ItemContainerContents.EMPTY);
+        } else {
+            safeComponents.set(CBCDataComponents.FUZE, ItemContainerContents.fromItems(List.of(fuze)));
+        }
     }
 
     @Override
@@ -83,13 +95,13 @@ public class FuzedBlockEntity extends BigCannonProjectileBlockEntity {
 	}
 
     public void setFuze(ItemStack itemStack) {
-        PatchedDataComponentMap components = new PatchedDataComponentMap(this.components());
+        PatchedDataComponentMap components = new PatchedDataComponentMap(DataComponentMap.EMPTY);
         if (itemStack.isEmpty()) {
             components.remove(CBCDataComponents.FUZE);
         } else {
             components.set(CBCDataComponents.FUZE, ItemContainerContents.fromItems(List.of(itemStack)));
         }
-        this.setComponents(components);
+        this.applyComponents(this.components(), components.asPatch());
     }
 
 	public boolean hasFuze() {